+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - The "in order" option is inverted, so it is now "in free order" and unchecked by default (but still `in_order` in the URL query param and in the API)
+- The checkbox of said option no longer gets disabled in Extended mode
 
 ## [9.5.1] - 2024-02-12